requirements.txt
# Requirements for multiphaseflowsr
# Install via conda : conda install --file requirements.txt

pytorch >= 1.11.0
sympy == 1.10.1
matplotlib
numpy <= 2.0.0
tqdm
pandas
pip
scikit-learn # For Feynman benchmark analysis script and for density in monitoring plots
jupyterlab # For running the notebooks

